+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   - Deployment detail - Clear log bugfix
   - Deployment dialog - Remove duplicated definition of price subscription
   - Deployment detail - Markdownified module description
+  - Improved responsive display of module overview
 
 ## [2.18.0] - 2021-05-06