+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 ### Changed
+
+  - Main - Detect unpaid subscription to show a more coherent message
   - Deployment - Bulk force delete support
   - Deployment - Bulk stop support
   - Fix visible console errors after update 
@@ -26,6 +28,7 @@
   - Deployment detail - Clear log bugfix
   - Deployment dialog - Remove duplicated definition of price subscription
   - Deployment detail - Markdownified module description
+  - Improved responsive display of module overview
 
 ## [2.18.0] - 2021-05-06