+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+  - Credential - Support gpg key
   - Main - Detect unpaid subscription to show a more coherent message
   - Deployment - Bulk force delete support
   - Deployment - Bulk stop support